Auth redirect fix, refactor employee spotlight, and dashboard weekday-trend + UI enhancements

+++ atp_streamlit/app.py
@@ -639,7 +639,7 @@ def get_employee_spotlight(conn, employee_id: int | None) -> dict | None:
                    COALESCE(e."Location", '') AS building,
                    GREATEST(0.0, ROUND(COALESCE((
                        SELECT SUM(ph.points) FROM points_history ph WHERE ph.employee_id = e.employee_id
-                   ), 0.0), 1)) AS point_total,
+                   ), 0.0)::numeric, 1)::float8) AS point_total,
                    (
                        SELECT MAX(ph2.point_date::date)
                          FROM points_history ph2
